feat: improved modularity/rustiness
