+++ src/avx2/mod.rs
@@ -1,0 +1,3 @@
+pub mod deser;
+pub mod stage1;
+pub mod utf8check;
